requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
autopep8==1.5.4
Django==3.1
django-admin==2.0.1
django-appconf==1.0.4
django-debug-toolbar==2.2
django-decouple==2.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.0.5
django-filter==2.3.0
django-graphql-auth==0.3.11
django-graphql-jwt==0.3.0
django-imagekit==4.0.2
django-phone-field==1.8.1
django-six==1.0.4
djangorestframework==3.11.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
pilkit==2.0
Pillow==7.2.0
ply==3.11
promise==2.3
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
Rx==1.6.1
screen==1.0.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
xlwt==1.3.0
